+++ NEWS.md
@@ -1,3 +1,5 @@
+# reqres (development version)
+
 # reqres 1.0.0
 
 * Use rlang native type checking instead of assertthat